Refactor getStorageType function in StorageAccessModule
The getStorageType function in the StorageAccessModule has been updated to handle null promises and return a string value. Previously, the function would fail if the promise was null. Now, the function will continue to run, returning 'unknown' in the case of an error and resolving the promise if one exists.