Refactor getSubdirectoryUri method for efficiency and correctness
The update refactors the getSubdirectoryUri method in ExternalStorageAccess.kt, replacing the previous approach of finding the file from the parent directory with directly constructing the URI. This simplifies the flow of the code, eliminates the situation of directory not found error, and increases the efficiency of method. The change was also reflected on fileSystem.ts, removing the deprecated announcement.