Refactor writeFile function in fileSystem module
The writeFile function has been altered to split the previous 'filePath' parameter into 'rootPath', 'filename', and 'extension'. This revision enhances flexibility and readability of the code by allowing separate inputs for different parts of a file. It also eliminates the need for internal splitting of file path, hence reducing error potential in filename and extension identification.